добавил обработчики done и delete
