+++ models.go
@@ -1,0 +1,17 @@
+package main
+
+type Task struct {
+	ID      string `json:"id"`
+	Date    string `json:"date"`
+	Title   string `json:"title"`
+	Comment string `json:"comment"`
+	Repeat  string `json:"repeat"`
+}
+
+type TasksResponse struct {
+	Tasks []Task `json:"tasks"`
+}
+
+var ErrorResponse struct {
+	Error string `json:"error,omitempty"`
+}
