+++ nextDateHandler.go
@@ -1,0 +1,33 @@
+package main
+
+import (
+	"net/http"
+	"time"
+)
+
+// Обработчик следующей даты
+func NextDateHandler(res http.ResponseWriter, req *http.Request) {
+	now := req.FormValue("now")
+	date := req.FormValue("date")
+	repeat := req.FormValue("repeat")
+
+	res.Header().Set("Content-Type", "application/json; charset=UTF-8")
+
+	nowTime, err := time.Parse(DateFormat, now)
+	if err != nil {
+		http.Error(res, "Некорректный формат даты", http.StatusBadRequest)
+		return
+	}
+	nextDate, err := NextDate(nowTime, date, repeat)
+	if err != nil {
+		http.Error(res, err.Error(), http.StatusBadRequest)
+		return
+	}
+	// Возвращаем ответ
+	_, err = res.Write([]byte(nextDate))
+	if err != nil {
+		http.Error(res, err.Error(), http.StatusBadRequest)
+		return
+	}
+
+}
