+++ tests/addtask_4_test.go
@@ -148,7 +148,6 @@ func TestAddTask(t *testing.T) {
 				continue
 			}
 			if today && task.Date != now.Format(`20060102`) {
-				fmt.Println(task.Date)
 				t.Errorf("Дата должна быть сегодняшняя %v", v)
 			}
 		}
